+++ src/sprite/animate.cpp
@@ -109,7 +109,7 @@ void			bunny_sprite_animate(t_bunny_sprite		*spr,
 		  || ++sprite.current_repeat >= anim->animation_repeat)
 		{
 		  sprite.current_repeat = 0;
-		  sprite.current_animation = anim->next_animation;
+		  sprite.current_animation = (int32_t)anim->next_animation;
 		  if (sprite.stop_repeat == false)
 		    {
 		      if (sprite.current_animation != -1)
